ecommerce-search-agent/pyproject.toml
[project]
name = "ecommerce-search-agent"
version = "0.1.0"
description = "An intelligent e-commerce search agent using LLMs and vector search"
authors = [
    {name = "Nasiat Hasan Fahim"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "chromadb>=0.4.15",
    "openai>=1.3.0",
    "langchain>=0.0.350",
    "langgraph>=0.0.10",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "tenacity>=8.2.3",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.10.0",
    "isort>=5.12.0",
    "mypy>=1.6.1",
    "ruff>=0.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []

[tool.hatch.build.targets.wheel]
packages = ["src"] 